Revert "add min and max dates"
This reverts commit 12af42bb321085a208cff8907c9f553f843f35c8.
