+++ src/views/schedule-actions/schedule-actions-menu/__tests__/schedule-actions-menu.test.tsx
@@ -7,7 +7,6 @@ import {
   type ScheduleActionsEnabledConfig,
 } from '@/config/dynamic/resolvers/schedule-actions-enabled.types';
 import { mockDescribeScheduleResponse } from '@/route-handlers/describe-schedule/__fixtures__/mock-describe-schedule-response';
-import { formatScheduleDetails } from '@/views/shared/hooks/use-describe-schedule/format-schedule-details';
 import mockResolvedConfigValues from '@/utils/config/__fixtures__/resolved-config-values';
 
 import { mockScheduleActionsConfig } from '../../__fixtures__/schedule-actions-config';
@@ -104,10 +103,10 @@ describe(ScheduleActionsMenu.name, () => {
 });
 
 function setup({
-  schedule = formatScheduleDetails(mockDescribeScheduleResponse),
+  schedule = mockDescribeScheduleResponse,
   actionsEnabledConfig,
 }: {
-  schedule?: ReturnType<typeof formatScheduleDetails> | undefined;
+  schedule?: typeof mockDescribeScheduleResponse | undefined;
   actionsEnabledConfig?: ScheduleActionsEnabledConfig;
 }) {
   const user = userEvent.setup();
