+++ src/views/schedule-actions/schedule-actions-menu/schedule-actions-menu.types.ts
@@ -1,10 +1,10 @@
 import { type ScheduleActionsEnabledConfig } from '@/config/dynamic/resolvers/schedule-actions-enabled.types';
-import { type FormattedScheduleDetails } from '@/views/shared/hooks/use-describe-schedule/use-describe-schedule.types';
+import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
 
 import { type SelectableScheduleAction } from '../config/schedule-actions.config';
 
 export type Props = {
-  schedule: FormattedScheduleDetails | undefined;
+  schedule: DescribeScheduleResponse | undefined;
   actionsEnabledConfig?: ScheduleActionsEnabledConfig;
   onActionSelect: (action: SelectableScheduleAction) => void;
 };
