+++ src/views/schedule-actions/schedule-actions-modal-content/schedule-actions-modal-content.tsx
@@ -154,7 +154,6 @@ export default function ScheduleActionsModalContent<
                 cluster={params.cluster}
                 domain={params.domain}
                 scheduleId={params.scheduleId}
-                schedule={schedule}
                 isSubmitted={isSubmitted}
               />
             )}
