+++ src/views/schedule-actions/schedule-actions-modal-content/schedule-actions-modal-content.types.ts
@@ -1,6 +1,6 @@
 import { type DefaultValues } from 'react-hook-form';
 
-import { type FormattedScheduleDetails } from '@/views/shared/hooks/use-describe-schedule/use-describe-schedule.types';
+import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
 
 import {
   type ScheduleAction,
@@ -10,7 +10,7 @@ import {
 export type Props<Result, FormData, SubmissionData> = {
   action: ScheduleAction<Result, FormData, SubmissionData>;
   params: ScheduleActionInputParams;
-  schedule?: FormattedScheduleDetails;
+  schedule?: DescribeScheduleResponse;
   onCloseModal: () => void;
   initialFormValues?: DefaultValues<FormData>;
 };
