+++ src/views/schedule-actions/schedule-actions-modal/schedule-actions-modal.types.ts
@@ -1,6 +1,6 @@
 import { type DefaultValues } from 'react-hook-form';
 
-import { type FormattedScheduleDetails } from '@/views/shared/hooks/use-describe-schedule/use-describe-schedule.types';
+import { type DescribeScheduleResponse } from '@/route-handlers/describe-schedule/describe-schedule.types';
 
 import { type ScheduleAction } from '../schedule-actions.types';
 
@@ -8,7 +8,7 @@ export type Props<Result, FormData, SubmissionData> = {
   domain: string;
   cluster: string;
   scheduleId: string;
-  schedule?: FormattedScheduleDetails;
+  schedule?: DescribeScheduleResponse;
   action: ScheduleAction<Result, FormData, SubmissionData> | undefined;
   onClose: () => void;
   initialFormValues?: DefaultValues<FormData>;
