+++ src/views/schedule-details/helpers/__tests__/format-schedule-details.test.ts
