+++ src/views/schedule-details/helpers/format-schedule-details.ts
@@ -1,14 +1,10 @@
-import { type PrettyJsonValue } from '@/components/pretty-json/pretty-json.types';
 import formatEnum from '@/utils/data-formatters/format-enum';
 import formatInputPayload from '@/utils/data-formatters/format-input-payload';
 import formatPayloadMap from '@/utils/data-formatters/format-payload-map';
 import formatRetryPolicy from '@/utils/data-formatters/format-retry-policy';
-import { type Props as TaskListLinkProps } from '@/views/shared/workflow-history-event-details-task-list-link/workflow-history-event-details-task-list-link.types';
+import { type DescribeScheduleResponse } from '@/views/shared/hooks/use-describe-schedule/use-describe-schedule.types';
 
-import {
-  type DescribeScheduleResponse,
-  type FormattedScheduleDetails,
-} from './use-describe-schedule.types';
+import { type FormattedScheduleDetails } from '../schedule-details.types';
 
 export function formatScheduleDetails(
   describeSchedule: DescribeScheduleResponse
