+++ src/views/shared/hooks/use-describe-schedule/__tests__/get-describe-schedule-query-options.test.ts
@@ -5,12 +5,11 @@ import {
   getMockRunningDescribeScheduleResponse,
 } from '@/route-handlers/describe-schedule/__fixtures__/mock-describe-schedule-response';
 import { type RequestError } from '@/utils/request/request-error';
-import { formatScheduleDetails } from '@/views/shared/hooks/use-describe-schedule/format-schedule-details';
 
 import getDescribeScheduleQueryOptions from '../get-describe-schedule-query-options';
 import {
   type DescribeScheduleQueryKey,
-  type FormattedScheduleDetails,
+  type DescribeScheduleResponse,
 } from '../use-describe-schedule.types';
 
 const params = {
@@ -26,9 +25,9 @@ const mockPausedDescribeScheduleResponse =
   getMockPausedDescribeScheduleResponse();
 
 type RefetchTypeIntervalArg = Query<
-  FormattedScheduleDetails,
+  DescribeScheduleResponse,
   RequestError,
-  FormattedScheduleDetails,
+  DescribeScheduleResponse,
   DescribeScheduleQueryKey
 >;
 
@@ -48,7 +47,7 @@ describe(getDescribeScheduleQueryOptions.name, () => {
 
     const mockQuery = {
       state: {
-        data: mockRunningDescribeScheduleResponse as FormattedScheduleDetails,
+        data: mockRunningDescribeScheduleResponse as DescribeScheduleResponse,
       },
     } as RefetchTypeIntervalArg;
 
@@ -65,7 +64,7 @@ describe(getDescribeScheduleQueryOptions.name, () => {
 
     const mockQuery = {
       state: {
-        data: mockPausedDescribeScheduleResponse as FormattedScheduleDetails,
+        data: mockPausedDescribeScheduleResponse as DescribeScheduleResponse,
       },
     } as RefetchTypeIntervalArg;
 
@@ -85,7 +84,7 @@ describe(getDescribeScheduleQueryOptions.name, () => {
 
     const mockQuery = {
       state: {
-        data: mockRunningDescribeScheduleResponse as FormattedScheduleDetails,
+        data: mockRunningDescribeScheduleResponse as DescribeScheduleResponse,
       },
     } as RefetchTypeIntervalArg;
 
@@ -108,48 +107,5 @@ describe(getDescribeScheduleQueryOptions.name, () => {
     });
 
     expect(options.throwOnError).toBe(true);
-  });
-
-  it('formats describe schedule response in queryFn', async () => {
-    const options = getDescribeScheduleQueryOptions(params);
-    const queryFn = options.queryFn;
-
-    if (!queryFn) {
-      throw new Error('Expected queryFn to be defined');
-    }
-
-    const describeSchedule = getMockRunningDescribeScheduleResponse({
-      action: {
-        startWorkflow: {
-          workflowType: { name: 'ScheduleWorker' },
-          taskList: null,
-          input: {
-            data: 'eyJ3b3JrZmxvd0FyZyI6InRlc3QtdmFsdWUifQ==',
-          },
-          workflowIdPrefix: 'schedule-prefix',
-          executionStartToCloseTimeout: null,
-          taskStartToCloseTimeout: null,
-          retryPolicy: null,
-          memo: null,
-          searchAttributes: null,
-        },
-      },
-    });
-
-    const fetchSpy = jest.spyOn(global, 'fetch').mockResolvedValue(
-      new Response(JSON.stringify(describeSchedule), {
-        status: 200,
-      })
-    );
-
-    const result = await queryFn({
-      queryKey: ['describeSchedule', params],
-      signal: new AbortController().signal,
-      meta: undefined,
-    });
-
-    expect(result).toEqual(formatScheduleDetails(describeSchedule));
-
-    fetchSpy.mockRestore();
   });
 });
